WebIDL/tests/submissions/heycam/object-001.html

<!DOCTYPE html>
<title>object</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.17 object
//
// When converting a JS value to an IDL object value, an exception is thrown
// for anything that is not an object.
//
// Since we're testing TrackEvent.track and TrackEventInit.track here which are
// of type "object?", we can't test object's handling of undefined and null.

var object = { };
var array = [];
var regexp = /abc/;

var values = [
  undefined,  null,      "undefined",
  null,       null,      "null",
  false,      undefined, "false",
  0,          undefined, "0",
  "",         undefined, '""',
  document,   document,  "document",
  object,     object,    "{ }",
  array,      array,     "[]",
  regexp,     regexp,    "/abc/"
];

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];

  test(function() {
    if (expected !== undefined) {
      var e = new TrackEvent("test", { track: value });
      assert_equals(e.track, expected);
    } else {
      assert_throws(new TypeError(), function() { new TrackEvent("test", { track: value }); });
    }
  }, "passing " + desc + " to the TrackEvent constructor correctly converts to an IDL object? value");
}
</script>
